+++ pom.xml
@@ -11,6 +11,14 @@
 
     <name>User Management WebApp</name>
 
+    <properties>
+        <!-- Specify the Java version to ensure compatibility -->
+        <java.version>1.8</java.version>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <maven.compiler.source>${java.version}</maven.compiler.source>
+        <maven.compiler.target>${java.version}</maven.compiler.target>
+    </properties>
+
     <dependencies>
         <!-- Servlet API Dependency -->
         <dependency>
@@ -34,24 +42,43 @@
             <version>2.8.9</version>
         </dependency>
 
+        <!-- JUnit Dependency for unit testing -->
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+            <version>4.13.2</version>
+            <scope>test</scope>
+        </dependency>
+
         <!-- Other dependencies as required -->
     </dependencies>
 
     <build>
         <finalName>user-management</finalName>
+
         <plugins>
+            <!-- This plugin is used for packaging WAR files -->
             <plugin>
-                <!-- This plugin is used for packaging WAR files -->
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-war-plugin</artifactId>
                 <version>3.3.1</version>
                 <configuration>
                     <!-- Specify the location of web.xml -->
-                    <webXml>WebContent/WEB-INF/web.xml</webXml>
+                    <!-- Not necessary if using Servlet 3.0+ with annotations -->
+                    <webXml>src/main/webapp/WEB-INF/web.xml</webXml>
                 </configuration>
             </plugin>
 
-            <!-- Other plugins as needed -->
+            <!-- Maven Compiler Plugin to specify Java version -->
+            <plugin>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <version>3.8.1</version>
+                <configuration>
+                    <source>${java.version}</source>
+                    <target>${java.version}</target>
+                </configuration>
+            </plugin>
         </plugins>
     </build>
+
 </project>
