+++ settings.xml
@@ -1,0 +1,7 @@
+<servers>
+    <server>
+        <id>TomcatServer</id>
+        <username>admin</username>
+        <password>1234</password>
+    </server>
+</servers>
