+++ src/main/webapp/WEB-INF/web.xml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<web-app xmlns="http://xmlns.jcp.org/xml/ns/javaee"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/javaee 
+         http://xmlns.jcp.org/xml/ns/javaee/web-app_4_0.xsd"
+         version="4.0">
+
+    <display-name>User Management Application</display-name>
+
+    <!-- Define the UserSearchServlet -->
+    <servlet>
+        <servlet-name>UserSearchServlet</servlet-name>
+        <servlet-class>com.example.servlets.UserSearchServlet</servlet-class>
+    </servlet>
+
+    <!-- Map the UserSearchServlet to /search -->
+    <servlet-mapping>
+        <servlet-name>UserSearchServlet</servlet-name>
+        <url-pattern>/search</url-pattern>
+    </servlet-mapping>
+
+    <!-- If you want to define a welcome file -->
+    <welcome-file-list>
+        <welcome-file>index.html</welcome-file>
+    </welcome-file-list>
+
+    <!-- Session configuration, error pages, etc. can also be defined here -->
+    
+</web-app>
